+++ Sstv.Host/Program.cs
@@ -4,9 +4,7 @@
 
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Resources;
-using Sstv.DomainExceptions;
 using Sstv.DomainExceptions.Extensions.DependencyInjection;
-using Sstv.DomainExceptions.Extensions.ProblemDetails;
 using Sstv.Host;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -22,48 +20,12 @@ builder.Services.AddOpenTelemetry()
         o.AddPrometheusExporter();
     });
 
-builder.Services.AddDomainExceptions(b =>
-{
-    b.WithErrorCodesDescriptionSource(FirstException.ErrorCodesDescriptionSource);
-    b.WithErrorCodesDescriptionSource(SecondErrorCodesException.ErrorCodesDescriptionSource);
-    b.WithErrorCodesDescriptionFromConfiguration();
-    b.UseDomainExceptionHandler();
-
-    b.ConfigureSettings = (sp, settings) =>
-    {
-        settings.GenerateExceptionIdAutomatically = true;     // default value
-        settings.CollectErrorCodesMetricAutomatically = true; // default value
-        settings.ThrowIfHasNoErrorCodeDescription = true;     // default value
-        settings.ErrorCodesDescriptionSource = null;          // manually set your own error description source instance
-        settings.DefaultErrorDescriptionProvider =            // override default error description func
-            errorCode => new ErrorDescription(errorCode, "N/A"); // default func
-
-        settings.OnErrorCreated += (error, _) =>
-        {
-            Console.WriteLine(error.ToString());
-        };
-    };
-});
+builder.Services.AddDomainException();
+builder.Services.AddProblemDetail();
 
 builder.Services
-    .AddProblemDetails(x =>
-    {
-        x.CustomizeProblemDetails = context =>
-        {
-            if (context.Exception is DomainException de)
-            {
-                context.ProblemDetails.Status = context.HttpContext.Response.StatusCode = ErrorCodeMapping.MapToStatusCode(de.GetDescription());
-            }
-            else
-            {
-                context.ProblemDetails = new ErrorCodeProblemDetails(ErrorCodes.Default.GetDescription())
-                {
-                    Status = context.HttpContext.Response.StatusCode = StatusCodes.Status500InternalServerError
-                };
-            }
-        };
-    })
     .AddControllers()
+    .AddValidationProblemDetails()
     .AddJsonOptions(options =>
     {
         options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
@@ -85,5 +47,3 @@ app.MapPrometheusScrapingEndpoint();
 app.UseErrorCodesDebugView();
 
 app.Run();
-
-