+++ Sstv.DomainExceptions.Extensions.SerilogEnricher/CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.0] - 2024-12-05
 
+### Changes
+
 - Update to .NET 9
 
 BREAKING CHANGES: